Split ChronoModel::Adapter test suite
